backend/requirements.txt
paddlepaddle==3.0.0
paddleocr==2.10.0
pillow==11.2.1
setuptools
fastapi
uvicorn
python-multipart
opencv-python-headless==4.11.0.86